Benchmark function and bug correction
A benchmarking option has been added to easily compile data about jobs and adjust the user variables.

An important bug that causes the temp files generated by the job to be created on the home folder on the clusters has been fixed. (added srun in the orca_job.sh template)